Revert the changes to use intermediate usize
